+++ changelog.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## [0.0.19]
 ### Changes
 - added the Galaxy Soul
 - fixed number formatting, now they comma seperate per 000 so 1,000 etc.